# CAB230 Server Side Application/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Movie API - n10869069",
        "description": "This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below.",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "n10869069@qut.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter JWT Bearer token **_only_**"
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/user/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "description": "Generates a new access token using a valid refresh token.",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "example": "ajsonwebtoken"
                                    }
                                },
                                "required": [
                                    "refreshToken"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token successfully refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "bearerToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "refreshToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "bearerToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Bearer",
                                        "expires_in": 600
                                    },
                                    "refreshToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Refresh",
                                        "expires_in": 86400
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid refresh request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ajsonwebtoken"
                                        }
                                    }
                                },
                                "example": {
                                    "error": true,
                                    "message": "Request body incomplete, refresh token required"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Click on 'Schema' below to see the possible error responses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "JWT token has expired"
                                        }
                                    }
                                }
                            }
                        }
                    }
                    
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Logs in a user and generates access and refresh tokens.",
                "operationId": "loginUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "mike@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "longExpiry": {
                                        "type": "bool",
                                        "example": false
                                    },
                                    "bearerExpiresInSeconds": {
                                        "type": "integer",
                                        "example": "600"
                                    },
                                    "refreshExpiresInSeconds": {
                                        "type": "integer",
                                        "example": "86400"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Log in successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "bearerToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "refreshToken": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string"
                                                },
                                                "expires_in": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "bearerToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Bearer",
                                        "expires_in": 600
                                    },
                                    "refreshToken": {
                                        "token": "ajsonwebtoken",
                                        "token_type": "Refresh",
                                        "expires_in": 86400
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid log in request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "error": true,
                                    "message": "Request body incomplete, both email and password are required"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Log in failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect email or password"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                          "application/json": {
                            "schema": {
                              "type": "object",
                              "properties": {
                                "Error": {
                                  "type": "boolean",
                                  "enum": [
                                    true
                                  ]
                                },
                                "Message": {
                                  "type": "string",
                                  "example": "Error in MySQL query"
                                }
                              }
                            }
                          }
                        }
                      }
                }
            }
        },
        "/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Invalidates the refresh token and logs out the user.",
                "operationId": "logoutUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "example": "ajsonwebtoken"
                                    }
                                },
                                "required": [
                                    "refreshToken"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token successfully invalidated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                false
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "error": false,
                                    "message": "Token successfully invalidated"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid refresh request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "error": true,
                                    "message": "Request body incomplete, refresh token required"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "JWT token has expired"
                                        }
                                    }
                                }
                            }
                        }
                    }
                    
                }
            }
        },
...
    }
}